Fix warnings

+++ examples/C++/Keys/keys.cpp
@@ -18,7 +18,7 @@
 using namespace eprosima::fastrtps;
 using namespace eprosima::fastrtps::rtps;
 
-constexpr int max_instances = 5;
+constexpr uint8_t max_instances = 5;
 
 //Enums and configuration structure
 enum Reliability_type
@@ -266,7 +266,7 @@ void keys()
             });
 
     // Registering 'max_instances' instances.
-    for (int i = 0; i < max_instances; i++)
+    for (uint8_t i = 0; i < max_instances; i++)
     {
         my_sample.key_value(i + 1);
         key_instances[i] = myPub->register_instance(&my_sample);
@@ -274,7 +274,7 @@ void keys()
 
     //Send 10 samples
     std::cout << "Publishing " << max_instances << " keys, 10 samples per key..." << std::endl;
-    for (int i = 0; i < max_instances; i++)
+    for (uint8_t i = 0; i < max_instances; i++)
     {
         for (int j = 0; j < 10; j++)
         {
@@ -304,7 +304,7 @@ void keys()
         sampleList.push_back(std::pair<int, int>(my_sample.index(), my_sample.key_value()));
     }
 
-    for (int key = 1; key <= max_instances; key++)
+    for (uint8_t key = 1; key <= max_instances; key++)
     {
         std::cout << "  On key " << std::to_string(key) << ": ";
         for (std::pair<int, int> n : sampleList)
@@ -319,7 +319,7 @@ void keys()
     std::cout << std::endl;
 
     std::cout << "Disposing all keys ..." << std::endl;
-    for (int i = 0; i < max_instances; i++)
+    for (uint8_t i = 0; i < max_instances; i++)
     {
         my_sample.key_value(i + 1);
         myPub->dispose(&my_sample, key_instances[i]);
@@ -331,7 +331,7 @@ void keys()
     {
         if (ChangeKind_t::NOT_ALIVE_DISPOSED == sample_info.sampleKind)
         {
-            for (int i = 0; i < max_instances; i++)
+            for (uint8_t i = 0; i < max_instances; i++)
             {
                 if (key_instances[i] == sample_info.iHandle)
                 {
@@ -364,7 +364,7 @@ void publisherKeys()
             });
 
     // Registering 'max_instances' instances.
-    for (int i = 0; i < max_instances; i++)
+    for (uint8_t i = 0; i < max_instances; i++)
     {
         my_sample.key_value(i + 1);
         myPub->register_instance(&my_sample);
@@ -372,7 +372,7 @@ void publisherKeys()
 
     //Send 10 samples
     std::cout << "Publishing " << max_instances << " keys, 10 samples per key..." << std::endl;
-    for (int i = 0; i < max_instances; i++)
+    for (uint8_t i = 0; i < max_instances; i++)
     {
         for (uint8_t j = 0; j < 10; j++)
         {
